pseudo code added
Pseudo code add for main portions of the ageclass module in LPJ. Header files with example data structures of ageclasses (stands) are included. Ancillary functions and data structures may be referenced, but may not be included in this pseudo code. As such, the pseudo code cannot be compiled or run. Ancillary functions, headers, and other files are necessary to fully compile LPJ. Nevertheless, these pseudo code files represent the foundation for the ageclass module and the code can be used to implement the method in other models.